bgy_admin/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
mysqlclient>=2.1.1
redis>=4.5.4
loguru>=0.7.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
django-filter>=23.2
+ gunicorn>=21.2.0
+ uvicorn>=0.24.0
+ python-dotenv>=1.0.0
+ drf-spectacular>=0.27.0